For a given x, you need to define the corresponding y where the vertical line first intercepts the polyline.
Or for a given y, you need to define the corresponding x where the horizontal line first intercepts the polyline.

The form should have a button labelled “Calculate”. When clicked, it should calculate the intercept value of the line constructed from the x,y coordinates listed in the label.

The calculation should read the x or y coordinate from the x or y text box, and write back the coordinate into the same text boxes.

If the x coordinate text box has a value and the y coordinate text box is empty, it should calculate the y coordinate, and write the value into the y coordinate text box.

If the y coordinate text box has a value and the x coordinate text box is empty, it should calculate the x coordinate, and write the value into the x coordinate text box.

If there are more than one possible x or y intercept, it should write back the first intercept that was found.